Cleanup README
- add installation section
- move low-level examples from README to examples dir
